Tidy up and remove old code
- remove old unused code
- fix doc build
