+++ src/FieldRecord.jl
@@ -14,7 +14,6 @@ struct FieldRecord{FieldData <: PB.AbstractData, Space <: PB.AbstractSpace, V, N
     data_dims::NTuple{N, PB.NamedDimension}
     mesh::Mesh
     attributes::Dict{Symbol, Any}
-    # coords_record::Vector{FixedCoord} # coordinates attached to record dimension
 end
 
 # create empty FieldRecord
@@ -201,21 +200,15 @@ any attached coordinates, for records and spatial region defined by `allselectar
   starting at the record with the nearest value of `fr.coords_record` before `first` and ending at the nearest record after
   `last`.
 
-Available `selectargs` depend on the grid `fr.mesh` and `fr.data_dims`, options include:
-- grid spatial dimensions 
-[- TODO grid spatial coordinates (if defined)]
-- `fr.data_dims`
+Available `selectargs` depend on the FieldRecord dimensions (as returned by `get_dimensions`, which will be a subset
+of grid spatial dimensions and Domain data dimensions) and corresponding attached coordinates (as returned by `get_coordinates`.
 
+Optional argument `coords` can be used to replace the attached coordinates for one or more dimensions.
+Format is a Vector of Pairs of `"<dim_name>"=>("<var_name1>", "<var_name2>", ...)`, 
+eg to replace a 1D column default pressure coordinate with a z coordinate:
 
-Optional argument `coords` can be used to supply coordinates from additional `FieldRecords`, replacing any coordinates
-attached to `fr`. Format is a Vector of Pairs of "dim_name"=>(cr1::FieldRecord, cr2::FieldRecord, ...)
+    coords=["cells"=>("atm.zmid", "atm.zlower", "atm.zupper")]
 
-Example: to replace a 1D column default pressure coordinate with a z coordinate:
-
-    coords=["z"=>(zmid::FieldRecord, zlower::FieldRecord, zupper::FieldRecord)]
-
-NB: the coordinates will be generated by applying `allselectargs`,
-so the supplied coordinate FieldRecords must have the same dimensionality as `fr`.
 """
 function get_array(
     fr::FieldRecord; 
@@ -231,163 +224,17 @@ function get_array(
     return get_array(fr, NamedTuple(allselectargs); coords=coords)
 end
 
-function get_array_orig(
-    fr::FieldRecord, allselectargs::NamedTuple; # allselectargs::NamedTuple=NamedTuple() creates a method ambiguity with deprecated form above
-    coords::Union{Nothing, AbstractVector}=nothing
-)
-
-    varray = nothing
-
-    try
-        # get data NB: with original coords or no coords
-        varray = _get_array(fr, allselectargs)
-
-        if !isnothing(coords)
-            check_coords_argument(coords) ||
-                error("coords argument should be of form 'coords=[\"z\"=>zmid::FieldRecord, zlower::FieldRecord, zupper::FieldRecord), ...]")
-
-            # get arrays for replacement coordinates
-            vec_coord_arrays = [
-                coord_name => Tuple(_get_array(coord_fr, allselectargs) for coord_fr in coord_fieldrecords) 
-                for (coord_name, coord_fieldrecords) in coords
-            ]
-
-            # replace coordinates
-            varray = update_coordinates(varray, vec_coord_arrays)
-        end
-    catch
-        frname = get(fr.attributes, :domain_name, "<none>")*"."*get(fr.attributes, :var_name, "<none>")
-        @warn "PALEOmodel.get_array(fr::FieldRecord) failed for variable $frname"
-        rethrow()
-    end
-
-    return varray
-end
-
-function _get_array(
-    fr::FieldRecord{FieldData, Space, V, N, Mesh, R}, allselectargs::NamedTuple, 
-) where {FieldData, Space, V, N, Mesh, R}
-
-    # select records to use and create PB.NamedDimension
-    ridx = 1:length(fr)
-    selectargs_region = NamedTuple()
-    selectargs_records = NamedTuple()
-    for (k, v) in zip(keys(allselectargs), values(allselectargs))
-        if k==:records
-            if v isa Integer
-                ridx = [v]                
-            else
-                ridx = v
-            end
-            selectargs_records = (records=v,)
-        elseif String(k) in fr.dataset.record_dim_coordinates
-            # find ridx corresponding to a coordinate
-            rfr = PB.get_field(fr.dataset, String(k))
-            ridx, cvalue = find_indices(rfr.records, v)
-            selectargs_records = NamedTuple((k=>cvalue,))    
-        else
-            selectargs_region = merge(selectargs_region, NamedTuple((k=>v,)))
-        end
-    end
-    records_coords = FixedCoord[]
-    for rcn in fr.dataset.record_dim_coordinates
-        rfr = PB.get_field(fr.dataset, rcn)
-        push!(records_coords, FixedCoord(rcn, rfr.records[ridx], rfr.attributes))
-    end
-    records_dim_coords = PB.NamedDimension(fr.dataset.record_dim.name, length(ridx)) => records_coords            
-
-    # add attributes for selection used
-    attributes = copy(fr.attributes)
-    isempty(selectargs_records) || (attributes[:filter_records] = selectargs_records;)
-    isempty(selectargs_region) || (attributes[:filter_region] = selectargs_region;)
-
-    # Generate name from attributes
-    name = default_fieldarray_name(attributes)
-
-    # Select selectargs_region 
-    if field_single_element(fr)
-        # create FieldArray directly from FieldRecord
-        data_dims_coords = get_data_dims_coords((fr.dataset, ridx), fr.data_dims)
-
-        isempty(selectargs_region) || 
-            throw(ArgumentError("invalid index $selectargs_region"))
-        if length(ridx) > 1
-            return FieldArray(
-                name,
-                fr.records[ridx], 
-                vcat(data_dims_coords, records_dim_coords),
-                attributes
-            )
-        else
-            # squeeze out records dimension
-            return FieldArray(
-                name,
-                fr.records[first(ridx)], 
-                data_dims_coords,
-                attributes
-            )
-        end
-    else
-        # pass through to Field
-
-        # get FieldArray from first Field record and use this to figure out array shapes etc
-        far = get_array(fr[first(ridx)], selectargs_region; coord_source=(fr.dataset, first(ridx)))
-        # TODO - Julia bug ? length(far.dims) returning wrong value, apparently triggered by this line
-        #        attributes = isnothing(attributes) ? Dict{Symbol, Any}() : copy(attributes)
-        # in get_array
-        if length(ridx) > 1
-            # add additional (last) dimension for records
-            favalues = Array{eltype(far.values), length(far.dims_coords)+1}(undef, size(far.values)..., length(ridx))
-            fa = FieldArray(
-                name,
-                favalues, 
-                vcat(far.dims_coords, records_dim_coords), 
-                attributes,
-            )
-            # copy values for first record
-            if isempty(far.dims_coords)
-                fa.values[1] = far.values
-            else
-                fa.values[fill(Colon(), length(far.dims_coords))..., 1] .= far.values
-            end
-        else
-            # squeeze out record dimension so this is just fa with attributes added
-            fa = FieldArray(
-                name, 
-                far.values,
-                far.dims_coords,
-                attributes
-            ) 
-        end
-        
-        # fill with values from FieldArrays for Fields for remaining records
-        if length(ridx) > 1
-            for (i, r) in enumerate(ridx[2:end])
-                far = get_array(fr[r], selectargs_region; coord_source=(fr.dataset, r))
-                if isempty(far.dims_coords)
-                    fa.values[i+1] = far.values
-                else
-                    fa.values[fill(Colon(), length(far.dims_coords))..., i+1] .= far.values
-                end
-            end
-        end
-
-        return fa
-    end
-
-end
-
 
 function get_array(
     fr::FieldRecord, allselectargs::NamedTuple; # allselectargs::NamedTuple=NamedTuple() creates a method ambiguity with deprecated form above
     coords=nothing,
     expand_cartesian=false, # can be overridden in allselectargs
     squeeze_all_single_dims=true, # can be overridden in allselectargs
-    verbose=true,
+    verbose=false,
 )
     frname = default_fieldarray_name(fr.attributes)
 
-    @info "get_array (begin): $frname, allselectargs $allselectargs"
+    verbose && @info "get_array (begin): $frname, allselectargs $allselectargs"
    
     if !isnothing(coords)
         check_coords_argument(coords) ||
@@ -561,7 +408,7 @@ function get_array(
     # Generate name from attributes, including selection suffixes
     name = default_fieldarray_name(attributes)
 
-    @info "get_array (end): $frname -> $name, allselectargs $allselectargs"
+    verbose && @info "get_array (end): $frname -> $name, allselectargs $allselectargs"
 
     return FieldArray(
         name,
@@ -570,6 +417,16 @@ function get_array(
         attributes,
     )
 end
+
+# check 'coords' of form [] or ["z"=>[ ... ], ] or ["z"=>(...),]
+check_coords_argument(coords) =
+    isa(coords, AbstractVector) && (
+        isempty(coords) || (
+            isa(coords, AbstractVector{<:Pair}) &&
+            isa(first(first(coords)), AbstractString) &&
+            isa(last(first(coords)), Union{AbstractVector, Tuple})
+        )
+    )
 
 # Read coordinates attached to dimension 'dim'
 # For record coordinate, ridx_to_use = nothing
