+++ src/PlotRecipes.jl
@@ -133,9 +133,6 @@ end
     heatmap(output::AbstractOutputWriter, var::AbstractString [, selectargs::NamedTuple] [; coords::AbstractVector])
     plot(outputs::Vector{<:AbstractOutputWriter}, vars::Union{AbstractString, Vector{<:AbstractString}} [, selectargs::NamedTuple] [; coords::AbstractVector])
 
-    plot(modeldata::AbstractModelData, vars::Union{AbstractString, Vector{<:AbstractString}} [, selectargs::NamedTuple] [; coords::AbstractVector])
-    heatmap(modeldata::AbstractModelData, var::AbstractString [, selectargs::NamedTuple] [; coords::AbstractVector])
-
 Plot recipe that calls `PB.get_field(output, var)`, and passes on to `plot(fr::FieldRecord, selectargs)`
 (see [`RecipesBase.apply_recipe(::Dict{Symbol, Any}, fr::FieldRecord, selectargs::NamedTuple)`](@ref))
 
@@ -152,8 +149,6 @@ Example: to replace a 1D column default pressure coordinate with a z coordinate:
  
     coords=["z"=>("atm.zmid", "atm.zlower", "atm.zupper")]
 
-NB: the coordinates will be generated by applying `selectargs`,
-so the supplied coordinate Variables must have the same dimensionality as `vars`.
 """
 RecipesBase.@recipe function f(
     output::AbstractOutputWriter,
@@ -166,16 +161,6 @@ RecipesBase.@recipe function f(
         vars = [vars]
     end
 
-    # if isnothing(coords)
-    #     coords_records=nothing
-    # else
-    #     check_coords_argument(coords) || 
-    #         error("coords argument should be of form 'coords=[\"z\"=>(\"atm.zmid\", \"atm.zlower\", \"atm.zupper\"), ...]")
-    #     coords_records = [
-    #         coord_name => Tuple(PB.get_field(output, cvn) for cvn in coord_varnames) 
-    #         for (coord_name, coord_varnames) in coords
-    #     ]
-    # end
     delete!(plotattributes, :coords)
             
     for var in vars
@@ -191,41 +176,6 @@ RecipesBase.@recipe function f(
         end
     end
 
-    return nothing
-end
-
-RecipesBase.@recipe function f(
-    modeldata::PB.AbstractModelData,
-    vars::Union{AbstractString, AbstractVector{<:AbstractString}},
-    selectargs::NamedTuple=NamedTuple();
-    coords=nothing, # NB: PlotRecipes doesn't support type annotation
-)
-    delete!(plotattributes, :coords)
-            
-    if isa(vars, AbstractString)
-        vars = [vars]
-    end
-
-    # broadcast any Vector-valued argument in selectargs
-    bcastargs = broadcast_dict([Dict{Symbol, Any}(pairs(selectargs))])
-
-    for var in vars        
-        # if var is of form <domain>.<name>.<structfield>, set :structfield to take a single field from struct-valued Var
-        var, var_structfield = _split_var_structfield(var)
-        if !isnothing(var_structfield)
-            structfield := var_structfield
-        end
-        
-        # broadcast selectargs
-        for sa in bcastargs
-            sa_nt = NamedTuple(sa)
-            RecipesBase.@series begin
-                # pass through to plot recipe for FieldArray
-                PALEOmodel.get_array(modeldata, var, sa_nt; coords=coords)
-            end
-        end
-    end
-  
     return nothing
 end
 
@@ -273,12 +223,13 @@ Calls [`get_array(fr, selectargs)`](@ref) and passes on to `plot(fa::FieldArray)
 
 Vector-valued fields in `selectargs` are "broadcasted" (generating a separate plot series for each combination)
 
-Optional argument `coords_records` can be used to supply plot coordinates from `FieldRecords`.
-Format is a Vector of Pairs of "coord_name"=>(cr1::FieldRecord, cr2::FieldRecord, ...)
-Example:
-    coords_records=["z"=>(zmid::FieldRecord, zlower::FieldRecord, zupper::FieldRecord)]
-to replace a 1D column default pressure coordinate with a z coordinate.   NB: the coordinates will be generated by applying `selectargs`,
-so the supplied coordinate FieldRecords must have the same dimensionality as `fr`.
+Optional argument `coords` can be used to supply plot coordinates from Variable in output.
+Format is a Vector of Pairs of "coord_name"=>("var_name1", "var_name2", ...)
+
+Example: to replace a 1D column default pressure coordinate with a z coordinate:
+ 
+    coords=["z"=>("atm.zmid", "atm.zlower", "atm.zupper")]
+
 """
 RecipesBase.@recipe function f(fr::FieldRecord, selectargs::NamedTuple, coords=nothing)
 
@@ -527,3 +478,84 @@ function broadcast_dict(dv::Vector{<:Dict})
         return dvout
     end
 end
+
+
+########################################################
+# Helper functions for output visualisation
+#########################################################
+
+"""
+    build_coords_edges(coords_vec::Vector{FixedCoord}) -> Vector{Float64}
+
+Build a vector of coordinate edges (length `n+1``) from `coords_vec`, assuming the PALEO
+convention that `coords_vec` contains three elements with 
+cell midpoints, lower edges, upper edges each of length `n`, in that order.
+
+Falls back to just returning the first entry in `coords_vec` for other cases.
+"""
+function build_coords_edges(dim::PB.NamedDimension, coords_vec::Vector{FixedCoord})
+
+    if isempty(coords_vec)
+        # no coordinate - just return indices
+        co_values = collect(1:(dim.size+1))
+        co_label = dim.name
+    elseif length(coords_vec) == 1 && is_boundary_coordinate(coords_vec[1])
+        co = coords_vec[1]
+        co_values = vcat(co.values[1, :], co.values[2, end])
+        co_label = append_units(co.name, co.attributes)
+    elseif length(coords_vec) == 1 || length(coords_vec) > 3
+        # 1 coordinate or something we don't understand - take first
+        co = first(coords_vec)
+        co_values = co.values
+        co_label = append_units(co.name, co.attributes)
+    elseif length(coords_vec) == 2 && is_boundary_coordinate(coords_vec[2])
+        # assume mid, boundary
+        co = coords_vec[2]
+        co_values = vcat(co.values[1, :], co.values[2, end])
+        co_label = append_units(co.name, co.attributes)
+    elseif length(coords_vec) in (2, 3)
+        # 2 coordinates assume lower, upper edges
+        # 3 coordinates assume mid, lower, upper
+        co_lower = coords_vec[end-1]
+        co_upper = coords_vec[end]
+        co_label = append_units(co_lower.name*", "*co_upper.name, co_lower.attributes)
+        first(co_lower.values) < first(co_upper.values) ||
+            @warn "build_coords_edges: $co_label co_lower is > co_upper - check model grid"            
+        if co_lower.values[end] > co_lower.values[1] # ascending order
+            co_lower.values[2:end] == co_upper.values[1:end-1] ||
+                @warn "build_coords_edges: $co_label lower and upper edges don't match"
+            co_values = [co_lower.values; co_upper.values[end]]
+        else # descending order
+            co_lower.values[1:end-1] == co_upper.values[2:end] ||
+                @warn "build_coords_edges: $co_label lower and upper edges don't match"
+            co_values = [co_upper.values[1]; co_lower.values]
+        end
+
+    end
+
+    return co_values, co_label
+end
+
+"guess coordinate edges from midpoints, assuming uniform spacing"
+function guess_coords_edges(x_midpoints)
+    first_x = x_midpoints[1] - 0.5*(x_midpoints[2] - x_midpoints[1])
+    last_x = x_midpoints[end] + 0.5*(x_midpoints[end] - x_midpoints[end-1])
+    return [first_x; 0.5.*(x_midpoints[1:end-1] .+ x_midpoints[2:end]); last_x]
+end
+
+
+"""
+    append_units(name::AbstractString, attributes) -> "name (units)"
+
+Utility function to append variable units string to a variable name for display.
+"""
+function append_units(name::AbstractString, attributes::Dict{Symbol, Any})
+    units = get(attributes, :units, "")
+    if isempty(units)
+        return name
+    else
+        return name*" ($units)"
+    end
+end
+
+append_units(name::AbstractString, attributes::Nothing) = name
